Quase finalizado

+++ Portal_NEABI/app/Http/Controllers/EventoController.php
@@ -4,6 +4,9 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Evento;
+use App\Models\User;
+use App\Models\User as eventoParticipante;
+use PhpParser\Node\Stmt\Foreach_;
 
 class EventoController extends Controller
 {
@@ -16,15 +19,24 @@ class EventoController extends Controller
     {
         $this->middleware('auth');
     }
-    
+    public function __invoke()
+    {
+        
+    }
     public function index()
     {
         //
-        $eventos = Evento::all();
+        $user = auth()->user();
+        if($user->userType == "admin"){
+        $eventos = $user->eventos;
         return view('evento.index',['eventos'=>$eventos]);
-        
-
+        }else{
+            return redirect('/')->with('msg',"você não tem permissão");
+        }  
+      
     }
+
+  
 
     /**
      * Show the form for creating a new resource.
@@ -34,7 +46,12 @@ class EventoController extends Controller
     public function create()
     {
         //
+        $user = auth()->user();
+        if($user->userType == "admin"){
         return view('evento.create');
+        }else{
+            return redirect('/')->with('msg',"você não tem permissão");
+        }
     }
 
     /**
@@ -46,20 +63,45 @@ class EventoController extends Controller
     public function store(Request $request)
     {
         //
-        $eventos = new Evento;
-        $eventos->nome = $request->nome ;
-        $eventos->descricao = $request->descricao ;
-        $eventos->data = $request->data;
-        $eventos->categoria = $request->categoria;
-        $eventos->url = $request->url;
-        $eventos->modo = $request->modo;
-        $eventos->organizadores = $request->organizadores;
-        $eventos->capacidade = $request->capacidade;
-        $eventos->hora_inicio = $request->hora_inicio;
-        $eventos->hora_termino = $request->hora_termino;
-        $eventos->status = true;
-        $eventos->save();
-       return view('evento.index',['eventos'=>$eventos]);
+        $request->validate([
+            'nome'=> 'required|min:5|max:100',
+            'descricao'=> 'required',
+            'data'=> 'required',
+             'modo' => 'required',
+             'organizadores'=> 'required',
+             'hora_inicio'=> 'required',
+             'hora_termino'=> 'required',
+             'capacidade'=> 'required',
+        ]);
+        $user = auth()->user();
+        if($user->userType =="admin"){
+        $evento = new Evento;
+        $evento->nome = $request->nome ;
+        $evento->descricao = $request->descricao ;
+        $evento->data = $request->data;
+        $evento->categoria = $request->categoria;
+        if($request->hasFile('url') && $request->file('url')->isValid()){
+            $requestImage = $request->url;
+            $extension =$requestImage->extension();
+            $imageName = md5($requestImage->getClientOriginalName() . strtotime("now")) . "." . $extension;
+            $requestImage->move(public_path('img/imagens'), $imageName);
+            $evento->url = $imageName;
+    
+            }
+        $evento->modo = $request->modo;
+        $evento->organizadores = $request->organizadores;
+        $evento->capacidade = $request->capacidade;
+        $evento->hora_inicio = $request->hora_inicio;
+        $evento->hora_termino = $request->hora_termino;
+        $evento->status = true;
+        $user=auth()->user();
+        $evento->user_id=$user->id;
+        $evento->save();
+        $eventos = $user->eventos;
+         return view('evento.index',['eventos'=>$eventos]);
+        }else{
+            return redirect('/')->with('msg',"você não tem permissão");
+        }
     }
 
     /**
@@ -70,9 +112,21 @@ class EventoController extends Controller
      */
     public function show($id)
     {
-        //
+     
+
+        $user = auth()->user();
+        $testuser = false;
+        if($user){
+            $userEventos = $user->eventosParticipante->toArray();
+            foreach($userEventos as $evento){
+                if($evento["id"] == $id){
+                    $testuser = true;
+                }
+            }
+        }
         $evento = Evento::findOrFail($id);
-        return view("evento.show", ['evento'=>$evento]);
+        //$eventoOwner = User::where('id',$evento->user_id)->first()->toArray();
+        return view("evento.show", ['evento'=>$evento,'testuser'=>$testuser]);
         
     }
 
@@ -85,8 +139,17 @@ class EventoController extends Controller
     public function edit($id)
     {
         //
-        $evento = Evento::findOrFail($id);
-        return view('evento.edit',['evento'=>$evento]);
+        $user = auth()->user();
+        if($user->userType == "admin"){
+                $evento = Evento::findOrFail($id);
+                if($user->id!=$evento->user_id){
+                    return redirect('/')->with('msg', 'Este Evento não é seu !');
+                }else{
+                return view('evento.edit',['evento'=>$evento]);
+                }
+        }else{
+            return redirect('/')->with('msg',"você não tem permissão");
+        }
         
     }
 
@@ -100,15 +163,42 @@ class EventoController extends Controller
     public function update(Request $request, $id)
     {
         {
-            $dadosform = $request->all();
+            $user = auth()->user();
+            //$dadosform = $request->all(); 
             $evento = Evento::findOrFail($id);
-            $update = $evento->update($dadosform);
-            if($update)
+             if($user->userType == "admin"){ 
+               // $update = $evento->update($dadosform);
+               /* if($update){
+                    return redirect()->route('evento.index');
+                } else {
+                    return redirect()->route('evento.edit', $id)->with('msg', 'Tente novamente !');
+                }*/
+                $evento->nome = $request->nome ;
+                $evento->descricao = $request->descricao ;
+                $evento->data = $request->data;
+                $evento->categoria = $request->categoria;
+                if($request->hasFile('url') && $request->file('url')->isValid()){
+                    $requestImage = $request->url;
+                    $extension =$requestImage->extension();
+                    $imageName = md5($requestImage->getClientOriginalName() . strtotime("now")) . "." . $extension;
+                    $requestImage->move(public_path('img/imagens'), $imageName);
+                    $evento->url = $imageName;
+    
+                }
+                $evento->modo = $request->modo;
+                $evento->organizadores = $request->organizadores;
+                $evento->capacidade = $request->capacidade;
+                $evento->hora_inicio = $request->hora_inicio;
+                $evento->hora_termino = $request->hora_termino;
+                $evento->status = $request->status;
+                $user=auth()->user();
+                $evento->user_id=$user->id;
+                $evento->update();
                 return redirect()->route('evento.index');
-            else 
-                return redirect()->route('evento.edit', $id);
-     
-         }
+                }else{
+                return redirect('/')->with('msg',"você não tem permissão");
+            }
+        }
         
     }
 
@@ -121,9 +211,50 @@ class EventoController extends Controller
     public function destroy($id)
     {
         //
-        Evento::findOrFail($id)->delete();
-        return redirect('evento');
+        $user = auth()->user();
+        $evento = Evento::findOrFail($id);
+        if($user->userType =="admin"){
+                if($user->id!=$evento->user_id){
+                    return redirect('/');
+                }else{
+                Evento::findOrFail($id)->delete();
+                return  redirect('evento')->with('msg',"Evento Deletado com Sucesso ");
+                }
+            }else{
+                return redirect('/');
+            }        
         
     }
    
+    public function participar($id){
+        $user = auth()->user();
+        $evento = Evento::findOrFail($id);
+        if(count($evento->users) < $evento->capacidade){
+        $user->eventosParticipante()->attach($id);
+        return redirect('/')->with('msg','Inscriçaõ feita com sucesso!');
+        }else{
+            return redirect('/')->with('msg','O Evento está Lotado!');
+        }
+    }
+
+    public function showEventos(){
+        $user = auth()->user();
+        $eventosParticipante = $user->eventosParticipante;
+        return view('evento.ShowEventos',['eventosParticipante'=>$eventosParticipante]);
+        
+    }
+    
+    public function cancelarInscricao($id){
+        $user = auth()->user();
+        $user->eventosParticipante()->detach($id);
+        return redirect('/')->with('msg', 'você cancelou sua incrição no Evento !');
+    }
+    public function nameUsers($id){
+        $user = auth()->user();
+        if($user->userType == "admin"){
+            $evento = Evento::findOrFail($id);
+            $NameUser = $evento->users;
+            return view('evento.usuarios',['NameUser'=>$NameUser]);
+        }
+    }
 }
