+++ Portal_NEABI/app/Http/Controllers/NoticiaController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 
 use App\Models\Noticia;
 use Illuminate\Support\Facades\Date;
+use App\Models\User;
 
 class NoticiaController extends Controller
 {
@@ -25,10 +26,16 @@ class NoticiaController extends Controller
     public function index()
     {
         //
-        $noticias = Noticia::all();
-       return view('noticia.index',['noticias'=>$noticias]);
+        $user =auth()->user();
+        if($user->userType == "admin"){
+        $noticias = $user->noticias;
+        return view('noticia.index',['noticias'=>$noticias]);
+        }else{
+            return redirect('/')->with('msg',"você não tem permissão");
+        }
     }
-
+    
+  
     /**
      * Show the form for creating a new resource.
      *
@@ -37,7 +44,12 @@ class NoticiaController extends Controller
     public function create()
     {
         //
+        $user = auth()->user();
+        if($user->userType == "admin"){
         return view('noticia.create');
+        }else{
+            return redirect('/')->with('msg',"você não tem permissão");
+        }
     }
 
     /**
@@ -49,14 +61,30 @@ class NoticiaController extends Controller
     public function store(Request $request)
     {
         //
+        $request->validate([
+            'titulo'=> 'required',
+            'descricao'=> 'required',
+            'categoria'=> 'required',
+        ]);
        $noticias = new Noticia;
        $noticias->titulo = $request->titulo ;
        $noticias->descricao = $request->descricao ;
        $noticias->categoria= $request->categoria ;
-       $noticias->url =$request->url;
+       if($request->hasFile('url') && $request->file('url')->isValid()){
+        $requestImage = $request->url;
+        $extension =$requestImage->extension();
+        $imageName = md5($requestImage->getClientOriginalName() . strtotime("now")) . "." . $extension;
+        $requestImage->move(public_path('img/imagens'), $imageName);
+        $noticias->url = $imageName;
+
+        }
        $noticias->data_edicao = new DateTime();
+       $user =auth()->user();
+       $noticias->user_id =$user->id;
        $noticias->save();
-       return view('noticia.index',['noticias'=>$noticias]);
+       $user = auth()->user();
+       $noticias = $user->noticias;
+       return view('noticia.index',['noticias'=>$noticias])->with('msg','Notícia Criada com Sucesso!');
     }
 
     /**
@@ -69,7 +97,8 @@ class NoticiaController extends Controller
     {
         //
         $noticia = Noticia::findOrFail($id);
-        return view("noticia.show", ['noticia'=>$noticia]);
+        $noticaOwner = User::where('id',$noticia->user_id)->first()->toArray();
+        return view("noticia.show", ['noticia'=>$noticia, 'noticiaOwner' =>$noticaOwner]);
        
     }
 
@@ -83,8 +112,16 @@ class NoticiaController extends Controller
     {
         //
         $noticias = Noticia::findOrFail($id);
-        return view('noticia.edit',['noticias'=>$noticias]);
-        
+        $user = auth()->user();
+        if($user->userType == "admin"){
+            if($user->id != $noticias->user_id){
+                return redirect('/');
+            }else{
+            return view('noticia.edit',['noticias'=>$noticias]);
+            }
+        }else{
+            return redirect('/')->with('msg',"você não tem permissão");
+        }
     }
 
     /**
@@ -96,15 +133,31 @@ class NoticiaController extends Controller
      */
     public function update(Request $request, $id)
     {
-       $dadosform = $request->all();
+       //$dadosform = $request->all();
        $noticia = Noticia::findOrFail($id);
-       $noticia->data_edicao = new DateTime();
-       $update = $noticia->update($dadosform);
-       if($update)
-           return redirect()->route('noticia.index');
-       else 
-           return redirect()->route('noticia.edit', $id);
+      // $update = $noticia->update($dadosform);
+       //if($update)
+         //  return redirect()->route('noticia.index')->with('msg','Notícia editada com Sucesso');
+       //else 
+          // return redirect()->route('noticia.edit', $id)->with('msg','Tente novamente!');
+       $noticia->titulo = $request->titulo ;
+       $noticia->descricao = $request->descricao ;
+       $noticia->categoria= $request->categoria ;
+       if($request->hasFile('url') && $request->file('url')->isValid()){
+        $requestImage = $request->url;
+        $extension =$requestImage->extension();
+        $imageName = md5($requestImage->getClientOriginalName() . strtotime("now")) . "." . $extension;
+        $requestImage->move(public_path('img/imagens'), $imageName);
+        $noticia->url = $imageName;
 
+        }
+       $noticia->data_edicao = new DateTime();
+       $user =auth()->user();
+       $noticia->user_id =$user->id;
+       $noticia->update();
+       $user = auth()->user();
+       $noticia = $user->noticias;
+       return redirect()->route('noticia.index')->with('msg','Notícia editada com Sucesso');
     }
 
     /**
@@ -116,9 +169,18 @@ class NoticiaController extends Controller
     public function destroy($id)
     {
         //
-        Noticia::findOrFail($id)->delete();
-        return redirect('noticia');
-        
+        $user = auth()->user();
+        if($user->userType == "admin"){
+            $noticia = Noticia::findOrFail($id);
+            if($user->id != $noticia->user_id){
+                return redirect('/');
+            }else{
+            Noticia::findOrFail($id)->delete();
+            return redirect('noticia')->with('msg','Noticia deletada com Sucesso!');
+            }
+        }else{
+            return redirect('/')->with('msg',"você não tem permissão");
+        }
     }
